add, edit & delete

+++ src/Customers.js
@@ -3,6 +3,7 @@ import { AgGridReact } from 'ag-grid-react';
 import { IconButton } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
 import AddCustomer from './AddCustomer';
+import EditCustomer from './EditCustomer';
 
 import 'ag-grid-community/styles/ag-grid.css';
 import 'ag-grid-community/styles/ag-theme-alpine.css';
@@ -26,13 +27,16 @@ function Customers () {
     }
 
     const deleteCustomer = (link) => {
-        console.log("deleted " + link[0].href);
-        fetch(link[0].href , {method: 'DELETE'})
-            .then(response => {
-                if(response.ok) {
-                    fetchCustomers();
+        if (window.confirm('Are you sure you want to delete customer?') == true) {
+            console.log("deleted " + link[0].href);
+            fetch(link[0].href , {method: 'DELETE'})
+                .then(response => {
+                    if(response.ok) {
+                        fetchCustomers();
+                    }
                 }
-            })
+            )
+        }
     }
     const saveCustomer = (customers) => {
         fetch("http://customerrest.herokuapp.com/api/customers", {
@@ -46,14 +50,28 @@ function Customers () {
           .catch((err) => console.error(err));
         console.log(JSON.stringify(customers));
     }
+
+    const saveEditedCustomer = (customer, link) => {
+        fetch(link[0].href , {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify(customer),
+        })
+          .then((response) => fetchCustomers())
+          .catch((err) => console.error(err));
+        console.log(JSON.stringify(customer));
+    }
+
     const [columnDefs, setColumnDefs] = useState([
-        {field: 'firstname', sortable: true, filter: true},
-        {field: 'lastname', sortable: true, filter: true},
+        {field: 'firstname', sortable: true, filter: true, width: 130},
+        {field: 'lastname', sortable: true, filter: true, width: 150},
         {field: 'streetaddress', sortable: true, filter: true},
         {field: 'postcode', sortable: true, filter: true, width: 120},
         {field: 'city', sortable: true, filter: true, width: 130},
         {field: 'email', sortable: true, filter: true},
-        {field: 'phone', sortable: true, filter: true},
+        {field: 'phone', sortable: true, filter: true, width: 170},
         {
             headerName: '',
             width: 80,
@@ -62,6 +80,13 @@ function Customers () {
             <IconButton color="error" onClick={() => deleteCustomer(params.value)}>
                 <DeleteIcon />
             </IconButton>
+        },
+        {
+            headerName: '',
+            width: 120,
+            field: 'links',
+            cellRenderer: params =>
+            <EditCustomer saveCustomer={saveEditedCustomer} link={params.value}/>
         }
     ]);
 
@@ -71,7 +96,7 @@ function Customers () {
             <div className="ag-theme-alpine" style={{height: '470px', width: '100%', margin: 'auto'}}>
                 <AgGridReact rowData={customers} columnDefs={columnDefs}
                     animateRows={true} rowSelection='multiple'
-                    paginationPageSize={10} pagination={true} />
+                    paginationPageSize={10} pagination={true} rowHeight={47} />
             </div>
         </>
     )
